pyproject.toml
[project]
name = "agent-chassis"
version = "0.1.0"
description = "A modular, scalable foundation for building AI agents with FastAPI, OpenAI, and MCP."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.1.0",
    "openai>=1.12.0",
    "httpx>=0.26.0",
    "mcp>=1.0.0",
    # Persistence (optional - for server-side session storage)
    "redis[hiredis]>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    # User Authentication (optional - OSP-14)
    "python-jose[cryptography]>=3.3.0",  # JWT tokens
    "bcrypt>=4.0.0",                     # Password hashing (direct, not via passlib)
    "aiosmtplib>=3.0.0",                 # Async SMTP for email
    "google-auth>=2.0.0",                # Google OAuth verification
    "email-validator>=2.0.0",            # Email validation in Pydantic
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "strict"
pythonpath = "."
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require external services like OpenAI API, Redis, PostgreSQL)",
]
# Note: Integration tests are excluded by default in CI/CD.
# To run integration tests locally: pytest -m integration
# To exclude them: pytest -m "not integration"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]